Added functionnalities to the settings UI
The settings UI can now be used to save the path leading to the folder containing his movies.

Added a Preferences class that will store data such as the movie folder path.

Added code that is now linked to the buttons in the UI

Added an error label that is shown when the user tries to save a path prior to choosing a folder containing his movies