Added queries execute functionnalities and changed DB fields
- I added a querie select method and a query no return method (insert into, update, delete) on CacheDb class

- I changed name and last_name fields of actor's and director's table  to only one name field.
It's very more simple for check if x or y person exists or no. Because it's difficult to know what's the name or last name of a person if he has 2 word with space. Which one is the name or last name ?
So, with one field "name" it's more simple.