Debugged some code and modified browseFiles evenement
First, I debugged some code on DBHandler class and add a method who get all the titles of movies on the DB.

Second, I changed the browseFiles evenement. When we click on the "scan directory" button, now, the real titles of all the films on your folder (so on the DB) is displayed.