Added check each 15sec internet connection functionnality
I added a functionnality who check each 15sec the internet connection. If there is a connection, a green signal appears and if there is no connection, a red signals appears. All this gunctionnality is an a Thread and run all long the application is open. So, in the main class I added a eventHandler who check if the application is closed. If we close it, we change a static variable in the controller class and the thread will stop. This way,  we can properly close the thread.