Fixed open movie file's linux bug 2 and Internet is enable method bug
I add some methods who check if we are on a Windows or Mac OS. If we are on one of the 2 (Mac or Windows), we can try to open a movie file with Desktop.open. Else, we are on linux and we display an error box for the moment, because Desktop.open doesn't work on Linux.

I also changed the InternetConnection.isEnable() method. Because with URL check, if we don't have internet, the method try more than 10sec to access to the URL. So now, I used InetAddress.isReachable. I can say to the method, wait maximum 500 ms. So now it work fine.