Added functionnality for hide the API key
I added the ApiHandler.getApiKey() method. It searches the value of the API Key on the .env file we have on our project file, but the .env file is not on the repository now. (.gitignore)

I also changed the web site we ping on the InternetConnection.isEnable() method. Before it was "8.8.8.8" now we have a real website who can respond correctly on the 80 port : www.google.com